Implemented partial config protection, and fixed double loading when command line was specified

+++ source/ftp_DELE.c
@@ -22,6 +22,12 @@ FTP_DECLARE(DELE) {
     if (build_path(session, session->cwd, args) != 0)
         return ftp_send_response(session, 553, "%s\r\n", strerror(errno));
 
+    /* return an error immediately if this is a config file. */
+    if (check_is_config(session->buffer, 1)) {
+        console_print(RED "Config '%s' overwrite rejected by policy.\n" RESET, session->buffer);
+        return ftp_send_response(session, 550, "delete was rejected by policy\r\n");
+    }
+
     /* try to unlink the path */
     rc = unlink(session->buffer);
     if (rc != 0) {
